fix(schedules): render end date server errors
